CRM_GestionAlumnos/backend/pyproject.toml
[project]
name = "crm-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "cload"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.27.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
